+++ src/database/migrations/2026_03_29_150300_create_staff_table.php
@@ -13,37 +13,37 @@ return new class extends Migration
     {
         Schema::create('staff', function (Blueprint $table) {
             $table->id();
-            
+
             // Basic information
             $table->string('name');
             $table->string('email')->unique();
             $table->string('phone');
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            
+
             // Position and role
             $table->enum('position', ['admin', 'manager', 'cook', 'delivery', 'cashier']);
             $table->text('responsibilities')->nullable(); // JSON of duties
-            
+
             // Employment details
             $table->date('hire_date');
-            $table->decimal('salary', 10, 2)->nullable();
+            $table->integer('salary')->nullable();
             $table->string('work_schedule')->nullable(); // JSON of working hours
             $table->boolean('is_active')->default(true);
-            
+
             // Performance tracking
             $table->integer('orders_processed')->default(0);
-            $table->decimal('total_tips', 10, 2)->default(0);
+            $table->integer('total_tips')->default(0);
             $table->float('rating')->nullable(); // Average customer rating
             $table->timestamp('last_login_at')->nullable();
-            
+
             // Emergency contact
             $table->string('emergency_contact_name')->nullable();
             $table->string('emergency_contact_phone')->nullable();
-            
+
             $table->rememberToken();
             $table->timestamps();
-            
+
             // Indexes
             $table->index('position');
             $table->index('is_active');
